+++ tests/test_package.py
@@ -11,8 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from __future__ import unicode_literals
-import platform
 from twine import package, exceptions
 
 import pretend
@@ -60,9 +58,7 @@ def test_sign_file_with_identity(monkeypatch):
 
 
 def test_run_gpg_raises_exception_if_no_gpgs(monkeypatch):
-    replaced_check_call = pretend.raiser(
-        package.FileNotFoundError('not found')
-    )
+    replaced_check_call = pretend.raiser(FileNotFoundError('not found'))
     monkeypatch.setattr(package.subprocess, 'check_call', replaced_check_call)
     gpg_args = ('gpg', '--detach-sign', '-a', 'pypircfile')
 
@@ -73,9 +69,7 @@ def test_run_gpg_raises_exception_if_no_gpgs(monkeypatch):
 
 
 def test_run_gpg_raises_exception_if_not_using_gpg(monkeypatch):
-    replaced_check_call = pretend.raiser(
-        package.FileNotFoundError('not found')
-    )
+    replaced_check_call = pretend.raiser(FileNotFoundError('not found'))
     monkeypatch.setattr(package.subprocess, 'check_call', replaced_check_call)
     gpg_args = ('not_gpg', '--detach-sign', '-a', 'pypircfile')
 
@@ -89,7 +83,7 @@ def test_run_gpg_falls_back_to_gpg2(monkeypatch):
 
     def check_call(arg_list):
         if arg_list[0] == 'gpg':
-            raise package.FileNotFoundError('gpg not found')
+            raise FileNotFoundError('gpg not found')
 
     replaced_check_call = pretend.call_recorder(check_call)
     monkeypatch.setattr(package.subprocess, 'check_call', replaced_check_call)
@@ -213,12 +207,6 @@ TWINE_1_5_0_WHEEL_HEXDIGEST = package.Hexdigest(
     'd86b0f33f0c7df49e888b11c43b417da5520cbdbce9f20618b1494b600061e67',
     'b657a4148d05bd0098c1d6d8cc4e14e766dbe93c3a5ab6723b969da27a87bac0',
 )
-
-if platform.python_implementation().lower() == 'pypy':
-    # pyblake2 refuses to install on PyPy
-    TWINE_1_5_0_WHEEL_HEXDIGEST = TWINE_1_5_0_WHEEL_HEXDIGEST._replace(
-        blake2=None,
-    )
 
 
 def test_hash_manager():