+++ twine/exceptions.py
@@ -107,3 +107,12 @@ class NonInteractive(TwineException):
     """Raised in non-interactive mode when credentials could not be found."""
 
     pass
+
+
+class InvalidPyPIUploadURL(TwineException):
+    """Repository configuration tries to use PyPI with an incorrect URL.
+
+    For example, https://pypi.org instead of https://upload.pypi.org/legacy.
+    """
+
+    pass