+++ docs/changelog.rst
@@ -3,6 +3,23 @@
 =========
 Changelog
 =========
+* :feature:`524`: Twine now unconditionally requires the keyring library
+  and no longer supports uninstalling ``keyring`` as a means to disable
+  that functionality. Instead, use ``keyring --disable`` keyring functionality
+  if necessary.
+* :feature:`518` Add Python 3.8 to classifiers.
+* :bug:`332` More robust handling of server response in ``--skip-existing``
+* :release:`2.0.0 <2019-09-24>`
+* :feature:`437` Twine now requires Python 3.6 or later. Use pip
+  9 or pin to "twine<2" to install twine on older Python versions.
+* :bug:`491` Require requests 2.20 or later to avoid reported security
+  vulnerabilities in earlier releases.
+* :release:`1.15.0 <2019-09-17>`
+* :feature:`488` Improved output on ``check`` command:
+  Prints a message when there are no distributions given to check.
+  Improved handling of errors in a distribution's markup, avoiding
+  messages flowing through to the next distribution's errors.
+* :release:`1.14.0 <2019-09-06>`
 * :feature:`456` Better error handling and gpg2 fallback if gpg not available.
 * :bug:`341` Fail more gracefully when encountering bad metadata
 * :feature:`459` Show Warehouse URL after uploading a package